Optimize TGA images that are RLE compressed
In the previous iteration, a run of pixels of the same color were
being naively recalculated. An optimized version of memset() has been
implemented to write 2-byte values starting a half-word or word
aligned address.

This gave us a speed increase from 12.0s to 0.456s.

Testing was done by counting the number of frames it took to draw the
image. There are 263 scanlines in a single frame. And each scanline
takes approximately 63.5us.

So, (frames * 263 * 63.5us) is the time it took.