Fix multiple issues
Fix the issue with timing. Since the INTBACK SMPC SMC was being called
at a specific scanline, there is an extremely limited amount of time
you're given before firing another interrupt for the subsequent
scanline. But because smpc_smc_call() was spinning until INTBACK
completes, it caused massive issues. Namely, the value SR was set to
was completely bogus. For peripheral collection, bit 7 wasn't set when
it should be.

Timing, I believe is still not correct as we can make use of the
optimized time feature.