Fix palette number pattern name config macros
The pattern name configuration macros for calculating the palette number
was wrong when > 16 colors was set.

If background is set to 16 colors and the pattern name is set at 1-word,
there are 128 16-color banks. The upper 3 bits are not set within the
pattern name data, but instead a VDP2 register. If the background is
greater than 16 colors, only 8 256-color banks are available and
addressable within the pattern name data (no need to set the VDP2
register).