Add collision detection in the physics component
Implemented the Separate-Axis Theorem (SAT) algorithm for collision
detection. Luckily, much of what is needed is pre-calculated as AABB's
are used. Implementing the general case for the algorithm requires quite
a bit calculations (a few dot products) before even performing an
intersection test. Considering the number of iterations needed per
collision, this would be pretty expensive.