Correctly calculate bitmap palette number
This was fairly confusing to deal with. There are 3 palette
color types. 16-colors, 256-colors, and 2,048-colors. Specifying a
palette number for 2,048-colors is invalid as it uses up nearly all
of CRAM (4 KiB). 256-color needs 3 bits as there are 8 possible
color banks. The weird case however is 16-colors. For 16-colors,
there are 128 possible banks, but only 8 are available (the lower 4-
bits of the palette number is 0).