Fix issues with sh-elf and m68k-elf not building
The odd issue is that binutils was incorrectly using the system's
(Cygwin) zlib instead of building its local version.

This proved to be a problem because the binutils executables would
require dynamically linking the system's DSO. In this case, on my
machine, it found:

C:\Program Files (x86)\GtkSharp\2.12\bin\zlib1.dll (x86)

The solution was to upgrade from binutils 2.24 to binutils 2.26.