Erase both VDP1 frame buffers
Aside from lack of 8BPP support, this should cover the different
resolutions the VDP2 would be set to.

This came from debugging a test and seeing a16px rectangle at the
bottom of garbage that wasn't being erased. This was due to the VDP1
only erasing from (0,0) to (319,223), but VDP2 set the resolution to
320x240.