Allow installing libyaul to a separate prefix
I personally found this very useful for packaging libyaul for distribution in
the Homebrew package manager, so that I can package libyaul itself separately
from its toolchain. It could also be useful for testing.

This adds a new PREFIX make variable. If not set, it defaults to INSTALL_ROOT;
that ensures it defaults to exactly the same way it worked before this commit.