Update sprite priority functions
There was a big misunderstanding with how sprite priority worked. It's
very confusing.

First, the frame buffer uses a sprite type 0 to 7 for a 16-bit
framebuffer, and sprite type 8 to F.

Each sprite type has a format used, and within that, there are a
number of priority bits. The priority bits are used as indices to the
values set by PRISA, PRISB, PRISC, and PRISD.