Fix flicker when in double-density interlace mode
The example vdp1-double-interlace exposed a flicker. This, I believe is
due to being too strict within the checks for interlace modes.

When in double-density interlace mode, who cares if we don't send a
command list to draw? We still need to change from even to odd (and
back) fields.