Fix issue w/ slave CPU raising interrupts
If there is a pending interrupt, SCU interrupts 0x41 and 0x43 are raised
even if both these interrupts are masked. The only way to disable the
interrupts from being raised is by setting the priority level to 14.

In the case where an interrupt does need to be raised, for example the
CPU-FRT ICI interrupt, then its priority level must be set to 15.