Fix issue with non-existent sections
The following sections are expected if -static is not passed to the
linker.

.interp:
This section holds the path name of a program interpreter. If the file
has a loadable segment that includes relocation, the sections' attributes
will include the SHF_ALLOC bit; otherwise, that bit will be off

.dynsym:
This section holds the dynamic linking symbol table

.dynstr:
This section holds strings needed for dynamic linking, most commonly the
strings that represent the names associated with symbol table entries

.hash:
This section holds a symbol hash table

.rela.*